Change project structure and makefile
