+++ components/global/observability-agent/pkg/adapter/adapter.go
