+++ components/global/observability-agent/pkg/adapter/testdata/adapter.crt
