+++ components/global/observability-agent/pkg/adapter/testdata/ca.pem
