+++ components/global/observability-agent/pkg/config/config_test.go
