+++ components/global/observability-agent/pkg/signals/signal.go
