+++ components/global/observability-agent/pkg/signals/signal_posix.go
