+++ components/global/observability-agent/pkg/signals/signal_windows.go
