+++ components/global/observability-agent/test/e2e/testdata/sample-operator-cfg.yaml
